requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
click==8.1.3
    # via -r requirements/base.in
pyyaml==6.0
    # via -r requirements/base.in
